deleted unnecessary pages
