--- src/pages/about.tsx
@@ -1,14 +1,0 @@
-// Step 1: Import React
-import * as React from "react";
-import Layout from "../components/layout";
-const AboutPage = () => {
-  return (
-    <Layout pageTitle="About Me">
-      <p>
-        Hi there! I'm the proud creator of this site, which I built with Gatsby.
-      </p>
-    </Layout>
-  );
-};
-// Step 3: Export your component
-export default AboutPage;
